Rename test files based on source code for Merge
`tests/sql/operators/test_agnostic_merge.py` -> `tests/sql/operators/test_merge.py`

- Tests in `test_sqlite_append.py` should already be covered by `test_append.py`
- `drop_table_postgres` was unused
- `astro.sqlite_utils` -> `astro.utils.sqlite_utils` - This function is temporary here until we refactor append and merge but in the meantime it was weird to have a file at top level
